Minor changes and fix mock server for e2e test
